refactor: Header, Chip, Tag, Tooltip, SvgIcon 컴포넌트 폴더구조 수정
