+++ src/components/header/HomeHeader.tsx
@@ -1,4 +1,4 @@
-import SVGIcon from '@/components/SvgIcon';
+import SvgIcon from '@/components/SvgIcon';
 import { useNavigate } from 'react-router';
 
 export default function HomeHeader() {
@@ -7,11 +7,11 @@ export default function HomeHeader() {
   return (
     <div className="flex h-full w-full items-center justify-between">
       <div className="flex items-center gap-[8px]">
-        <SVGIcon width={28} height={36} name="HeaderLogo" />
-        <SVGIcon width={72} height={32} name="HeaderTypo" />
+        <SvgIcon width={28} height={36} name="HeaderLogo" />
+        <SvgIcon width={72} height={32} name="HeaderTypo" />
       </div>
       <div className="flex max-w-[72px] items-center gap-[8px]">
-        <SVGIcon
+        <SvgIcon
           className="mr-2"
           width={32}
           height={32}
@@ -19,7 +19,7 @@ export default function HomeHeader() {
           active={false}
           onClick={() => navigate('/notifications')}
         />
-        <SVGIcon width={32} height={32} name="Search" />
+        <SvgIcon width={32} height={32} name="Search" />
       </div>
     </div>
   );
