--- src/components/tooltip/types.ts
@@ -1,15 +1,0 @@
-export type TooltipProps = {
-  size: 'small' | 'medium';
-  direction: TooltipDirections;
-  message: string;
-  isBubble: boolean;
-  className?: string;
-};
-
-export type TooltipDirections =
-  | 'topLeft'
-  | 'topMiddle'
-  | 'topRight'
-  | 'bottomLeft'
-  | 'bottomMiddle'
-  | 'bottomRight';
